pyproject.toml
[project]
name = "ad-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Andrew Dorton", email = "addorton@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "azure-storage-blob>=12.19.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
]

[tool.setuptools]
packages = ["ad_pipeline"]

[tool.uv]
package = true
